i#1703 cache simulator: speed up replacement via smaller counters
Shrinks the per-line counter from 64 bits to 32 bits to improve performance
on fib by 4%.  The reduced granularity should still be sufficient for
reasonable LRU or LFU.

Review-URL: https://codereview.appspot.com/248240043

+++ clients/drcachesim/simulator/cache_line.h
@@ -46,7 +46,11 @@ class cache_line_t
 
     addr_t tag;
     bool valid;
-    int_least64_t counter; // for use by replacement policies
+
+    // XXX: using int_least64_t here results in a ~4% slowdown for 32-bit apps.
+    // A 32-bit counter should be sufficient but we may want to revisit.
+    // We already have inttypes.h so we can reinstate int_least64_t easily.
+    int counter; // for use by replacement policies
 };
 
 #endif /* _CACHE_LINE_H_ */
